# docs/swagger.yaml
basePath: /api
definitions:
...
  entities.Booking:
    properties:
      created_at:
        type: string
      days:
        type: integer
      id:
        type: integer
      room_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      vender_id:
        type: integer
    type: object
  entities.JSONResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
...
  entities.RoomPayload:
    properties:
      cost:
        type: string
      status:
        type: string
      vendor:
        type: integer
    type: object
...
info:
  contact:
    email: bicosteve4@gmail.com
    name: Bico Oloo
    url: https://github.com/bicosteve
  description: Booking API to make booking reservations
  title: Booking API
  version: "1.0"
paths:
  /api/admin/book/{booking_id}/{room_id}:
    delete:
      consumes:
      - application/json
      description: deletes a booking
      operationId: delete-booking
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bookings not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: update user booking
      tags:
      - bookings
  /api/admin/book/all:
    get:
      consumes:
      - application/json
      description: Retrieves all booking
      operationId: admin-bookings
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bookings not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: get all bookings for admin user
      tags:
      - bookings
...
  /api/user/{room_i}:
    get:
      consumes:
      - application/json
      description: Receives room_id then retrieves a booking
      operationId: get-booking
      parameters:
      - description: To get a room
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/entities.Booking'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: get a booking
      tags:
      - bookings
  /api/user/all:
    get:
      consumes:
      - application/json
      description: Receives room_id then retrieves a booking
      operationId: user-bookings
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bookings not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: get all bookings
      tags:
      - bookings
  /api/user/book:
    post:
      consumes:
      - application/json
      description: Receives booking payload, validates it, create a booking
      operationId: create-booking
      parameters:
      - description: Create room
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.RoomPayload'
      produces:
      - application/json
      responses:
        "201":
          description: '{"msg":"created"}'
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: user create a booking
      tags:
      - bookings
  /api/user/book/{booking_id}:
    put:
      consumes:
      - application/json
      description: Updates a booking
      operationId: update-booking
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bookings not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: update user booking
      tags:
      - bookings
...
  /api/user/verify/{room_id}:
    get:
      consumes:
      - application/json
      description: Receives room_id, validates it then confirm booking
      operationId: verify-booking
      parameters:
      - description: To verify room
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking success
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: user verify booking
      tags:
      - bookings
schemes:
- http
swagger: "2.0"